feat: add exception handling and implement Category CRUD operations

+++ src/main/java/com/rishika/product_category_api/service/ProductService.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface ProductService {
     Product createProduct(ProductRequestDTO dto);
-    Product getProductById(Integer id);
+    Product getProductById(Long id);
     List<Product> getAllProducts();
-     void deleteProduct(Integer id);
+     void deleteProduct(Long id);
 }
